+++ src/dipdup/indexes/evm_subsquid.py
@@ -11,6 +11,7 @@ from web3 import Web3
 from dipdup.config import EvmSubsquidIndexConfigU
 from dipdup.config.evm import EvmContractConfig
 from dipdup.config.evm_node import EvmNodeDatasourceConfig
+from dipdup.config.evm_subsquid import EvmSubsquidDatasourceConfig
 from dipdup.context import DipDupContext
 from dipdup.datasources import IndexDatasource
 from dipdup.datasources.evm_node import NODE_LAST_MILE
@@ -27,7 +28,7 @@ from dipdup.prometheus import Metrics
 SUBSQUID_READAHEAD_LIMIT = 10000
 
 IndexConfigT = TypeVar('IndexConfigT', bound=EvmSubsquidIndexConfigU)
-DatasourceT = TypeVar('DatasourceT', bound=EvmSubsquidDatasource)
+DatasourceT = TypeVar('DatasourceT', bound=EvmSubsquidDatasource | EvmNodeDatasource)
 
 
 _sighashes: dict[str, str] = {}
@@ -62,14 +63,19 @@ class SubsquidIndex(
     ) -> None:
         super().__init__(ctx, config, datasource)
 
-        node_field = self._config.datasource.node
-        if node_field is None:
-            node_field = ()
-        elif isinstance(node_field, EvmNodeDatasourceConfig):
-            node_field = (node_field,)
-        self._node_datasources = tuple(
-            self._ctx.get_evm_node_datasource(node_config.name) for node_config in node_field
-        )
+        if isinstance(datasource, EvmSubsquidDatasource) and isinstance(config.datasource, EvmSubsquidDatasourceConfig):
+            node_field = config.datasource.node
+            if node_field is None:
+                node_field = ()
+            elif isinstance(node_field, EvmNodeDatasourceConfig):
+                node_field = (node_field,)
+            self._node_datasources = tuple(
+                self._ctx.get_evm_node_datasource(node_config.name) for node_config in node_field
+            )
+        elif isinstance(datasource, EvmNodeDatasource) and isinstance(config.datasource, EvmNodeDatasourceConfig):
+            self._node_datasources = (datasource,)
+        else:
+            raise FrameworkException('Invalid datasource type')
 
     @abstractmethod
     async def _synchronize_subsquid(self, sync_level: int) -> None: ...
@@ -118,9 +124,6 @@ class SubsquidIndex(
         self._logger.info('Subsquid is %s levels behind; %s available', subsquid_lag, subsquid_available)
         if subsquid_available < NODE_LAST_MILE:
             return node_sync_level
-        if self._config.node_only:
-            self._logger.debug('`node_only` flag is set; using node anyway')
-            return node_sync_level
         return None
 
     async def _synchronize(self, sync_level: int) -> None:
@@ -133,8 +136,12 @@ class SubsquidIndex(
         if levels_left <= 0:
             return
 
-        subsquid_sync_level = await self.datasource.get_head_level()
-        Metrics.set_sqd_processor_chain_height(subsquid_sync_level)
+        if isinstance(self.datasource, EvmSubsquidDatasource):
+            subsquid_sync_level = await self.datasource.get_head_level()
+            Metrics.set_sqd_processor_chain_height(subsquid_sync_level)
+        else:
+            subsquid_sync_level = 0
+
         node_sync_level = await self._get_node_sync_level(subsquid_sync_level, index_level)
 
         # NOTE: Fetch last blocks from node if there are not enough realtime messages in queue